Update HangmanGUI (startGame() method)
In the startGame() method, I retrieve the username and level given by the user from the appropriate text fields. Then I build a new User object with the specified username and add it to the GameManager. Next, I utilise the GameManager to start a new game for the current user at the desired level.